Move functions to shared.
